pyproject.toml
[project]
name = "offense-bias-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aif360>=0.6.1",
    "anthropic>=0.53.0",
    "fastparquet>=2024.11.0",
    "google-generativeai>=0.8.5",
    "pandas>=2.3.0",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
    "torch>=2.7.1",
    "transformers>=4.52.4",
]
